fix(ui): stop resuming restored auth attempts
Align AuthView cold-start behavior with the iOS and JavaScript SDKs by removing global sign-in and sign-up attempt routing. Clear navigation restored after process death while preserving same-process configuration changes.
