+++ source/ui/src/main/java/com/clerk/ui/auth/AuthStartView.kt
@@ -233,7 +233,6 @@ internal fun AuthStartViewImpl(
                 } else {
                   authState.authStartIdentifier
                 }
-              authState.enableInProgressAuthAttemptResume()
               authStartViewModel.startAuth(
                 authMode = authState.mode,
                 isPhoneNumberFieldActive = phoneActive,
@@ -289,7 +288,6 @@ internal fun AuthStartViewImpl(
                 provider = lastUsedSocialProvider,
                 modifier = Modifier.fillMaxWidth(),
                 onClick = {
-                  authState.enableInProgressAuthAttemptResume()
                   authStartViewModel.authenticateWithSocialProvider(
                     provider = it,
                     transferable = authState.mode.transferable,
@@ -307,7 +305,6 @@ internal fun AuthStartViewImpl(
             ClerkSocialRow(
               providers = socialProvidersMinusLastUsed.toImmutableList(),
               onClick = {
-                authState.enableInProgressAuthAttemptResume()
                 authStartViewModel.authenticateWithSocialProvider(
                   provider = it,
                   transferable = authState.mode.transferable,
