--- source/ui/src/test/java/com/clerk/ui/auth/AuthViewPendingAuthRoutingTest.kt
@@ -1,330 +1,0 @@
-package com.clerk.ui.auth
-
-import android.content.Context
-import androidx.navigation3.runtime.NavBackStack
-import androidx.navigation3.runtime.NavKey
-import androidx.test.core.app.ApplicationProvider
-import com.clerk.api.Constants
-import com.clerk.api.network.model.factor.Factor
-import com.clerk.api.network.model.verification.Verification
-import com.clerk.api.signin.SignIn
-import com.clerk.api.signup.SignUp
-import io.mockk.every
-import io.mockk.mockk
-import io.mockk.verify
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertTrue
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
-
-@RunWith(RobolectricTestRunner::class)
-class AuthViewPendingAuthRoutingTest {
-
-  private lateinit var context: Context
-  private val backStack = mockk<NavBackStack<NavKey>>(relaxed = true)
-  private lateinit var authState: AuthState
-
-  @Before
-  fun setUp() {
-    context = ApplicationProvider.getApplicationContext()
-    preferences().edit().clear().commit()
-    every { backStack.add(any()) } returns true
-    authState =
-      AuthState(
-        mode = AuthMode.SignInOrUp,
-        backStack = backStack,
-        sharedPreferences = preferences(),
-      )
-  }
-
-  @Test
-  fun `resumeInProgressAuthAttempt routes sign in to second factor from auth start`() {
-    val factor = Factor(strategy = Constants.Strategy.TOTP)
-    val signIn =
-      SignIn(
-        id = "sign_in_123",
-        status = SignIn.Status.NEEDS_SECOND_FACTOR,
-        supportedSecondFactors = listOf(factor),
-      )
-
-    resumeInProgressAuthAttempt(
-      authState = authState,
-      top = AuthDestination.AuthStart,
-      signIn = signIn,
-      signUp = null,
-      onAuthComplete = {},
-    )
-
-    verify(exactly = 1) { backStack.add(AuthDestination.SignInFactorTwo(factor = factor)) }
-  }
-
-  @Test
-  fun `discardRestoredAuthNavigation clears restored reset factor stack`() {
-    val resetFactor = Factor(strategy = Constants.Strategy.RESET_PASSWORD_EMAIL_CODE)
-    val restoredBackStack =
-      NavBackStack<NavKey>(
-        AuthDestination.AuthStart,
-        AuthDestination.SignInForgotPassword,
-        AuthDestination.SignInFactorOne(resetFactor),
-      )
-
-    discardRestoredAuthNavigation(backStack = restoredBackStack, resumeInProgressAuthAttempt = true)
-
-    assertEquals(listOf(AuthDestination.AuthStart), restoredBackStack.toList())
-  }
-
-  @Test
-  fun `discardRestoredAuthNavigation clears restored new password stack`() {
-    val restoredBackStack =
-      NavBackStack<NavKey>(AuthDestination.AuthStart, AuthDestination.SignInSetNewPassword)
-
-    discardRestoredAuthNavigation(backStack = restoredBackStack, resumeInProgressAuthAttempt = true)
-
-    assertEquals(listOf(AuthDestination.AuthStart), restoredBackStack.toList())
-  }
-
-  @Test
-  fun `discardRestoredAuthNavigation preserves ordinary auth stack when resume is enabled`() {
-    val factor = Factor(strategy = Constants.Strategy.TOTP)
-    val restoredBackStack =
-      NavBackStack<NavKey>(AuthDestination.AuthStart, AuthDestination.SignInFactorTwo(factor))
-
-    discardRestoredAuthNavigation(backStack = restoredBackStack, resumeInProgressAuthAttempt = true)
-
-    assertEquals(
-      listOf(AuthDestination.AuthStart, AuthDestination.SignInFactorTwo(factor)),
-      restoredBackStack.toList(),
-    )
-  }
-
-  @Test
-  fun `discardRestoredAuthNavigation clears ordinary auth stack when resume is disabled`() {
-    val factor = Factor(strategy = Constants.Strategy.TOTP)
-    val restoredBackStack =
-      NavBackStack<NavKey>(AuthDestination.AuthStart, AuthDestination.SignInFactorTwo(factor))
-
-    discardRestoredAuthNavigation(
-      backStack = restoredBackStack,
-      resumeInProgressAuthAttempt = false,
-    )
-
-    assertEquals(listOf(AuthDestination.AuthStart), restoredBackStack.toList())
-  }
-
-  @Test
-  fun `resumeInProgressAuthAttempt respects host opt out`() {
-    val factor = Factor(strategy = Constants.Strategy.TOTP)
-    val optedOutAuthState =
-      AuthState(
-        mode = AuthMode.SignInOrUp,
-        backStack = backStack,
-        sharedPreferences = preferences(),
-        resumeInProgressAuthAttempt = false,
-      )
-    val signIn =
-      SignIn(
-        id = "sign_in_123",
-        status = SignIn.Status.NEEDS_SECOND_FACTOR,
-        supportedSecondFactors = listOf(factor),
-      )
-
-    resumeInProgressAuthAttempt(
-      authState = optedOutAuthState,
-      top = AuthDestination.AuthStart,
-      signIn = signIn,
-      signUp = null,
-      onAuthComplete = {},
-    )
-
-    verify(exactly = 0) { backStack.add(any()) }
-  }
-
-  @Test
-  fun `resumeInProgressAuthAttempt does not restore reset password code screen`() {
-    val resetFactor = Factor(strategy = Constants.Strategy.RESET_PASSWORD_EMAIL_CODE)
-    val signIn =
-      SignIn(
-        id = "sign_in_123",
-        status = SignIn.Status.NEEDS_FIRST_FACTOR,
-        supportedFirstFactors = listOf(resetFactor),
-        firstFactorVerification =
-          Verification(
-            status = Verification.Status.UNVERIFIED,
-            strategy = Constants.Strategy.RESET_PASSWORD_EMAIL_CODE,
-          ),
-      )
-
-    resumeInProgressAuthAttempt(
-      authState = authState,
-      top = AuthDestination.AuthStart,
-      signIn = signIn,
-      signUp = null,
-      onAuthComplete = {},
-    )
-
-    verify(exactly = 0) { backStack.add(any()) }
-  }
-
-  @Test
-  fun `resumeInProgressAuthAttempt does not restore new password screen`() {
-    val signIn = SignIn(id = "sign_in_123", status = SignIn.Status.NEEDS_NEW_PASSWORD)
-
-    resumeInProgressAuthAttempt(
-      authState = authState,
-      top = AuthDestination.AuthStart,
-      signIn = signIn,
-      signUp = null,
-      onAuthComplete = {},
-    )
-
-    verify(exactly = 0) { backStack.add(any()) }
-  }
-
-  @Test
-  fun `resumeInProgressAuthAttempt does not reroute when auth stack is already past root`() {
-    val factor = Factor(strategy = Constants.Strategy.TOTP)
-    val signIn =
-      SignIn(
-        id = "sign_in_123",
-        status = SignIn.Status.NEEDS_SECOND_FACTOR,
-        supportedSecondFactors = listOf(factor),
-      )
-
-    resumeInProgressAuthAttempt(
-      authState = authState,
-      top = AuthDestination.SignInFactorTwo(factor),
-      signIn = signIn,
-      signUp = null,
-      onAuthComplete = {},
-    )
-
-    verify(exactly = 0) { backStack.add(any()) }
-  }
-
-  @Test
-  fun `resumeInProgressAuthAttempt does not reroute after identifier edit returns to root`() {
-    val factor = Factor(strategy = Constants.Strategy.TOTP)
-    val signIn =
-      SignIn(
-        id = "sign_in_123",
-        status = SignIn.Status.NEEDS_SECOND_FACTOR,
-        supportedSecondFactors = listOf(factor),
-      )
-
-    authState.navigateToAuthStartForIdentifierEdit()
-
-    resumeInProgressAuthAttempt(
-      authState = authState,
-      top = AuthDestination.AuthStart,
-      signIn = signIn,
-      signUp = null,
-      onAuthComplete = {},
-    )
-
-    verify(exactly = 0) { backStack.add(any()) }
-  }
-
-  @Test
-  fun `resumeInProgressAuthAttempt completes auth after identifier edit returns to root`() {
-    val signIn = SignIn(id = "sign_in_123", status = SignIn.Status.COMPLETE)
-    var completed = false
-
-    authState.navigateToAuthStartForIdentifierEdit()
-
-    resumeInProgressAuthAttempt(
-      authState = authState,
-      top = AuthDestination.AuthStart,
-      signIn = signIn,
-      signUp = null,
-      onAuthComplete = { completed = true },
-    )
-
-    assertTrue(completed)
-    verify(exactly = 0) { backStack.add(any()) }
-  }
-
-  @Test
-  fun `resumeInProgressAuthAttempt completes reset auth after host opt out`() {
-    val optedOutAuthState =
-      AuthState(
-        mode = AuthMode.SignInOrUp,
-        backStack = backStack,
-        sharedPreferences = preferences(),
-        resumeInProgressAuthAttempt = false,
-      )
-    val signIn =
-      SignIn(
-        id = "sign_in_123",
-        status = SignIn.Status.COMPLETE,
-        firstFactorVerification =
-          Verification(
-            status = Verification.Status.VERIFIED,
-            strategy = Constants.Strategy.RESET_PASSWORD_EMAIL_CODE,
-          ),
-      )
-    var completed = false
-
-    resumeInProgressAuthAttempt(
-      authState = optedOutAuthState,
-      top = AuthDestination.AuthStart,
-      signIn = signIn,
-      signUp = null,
-      onAuthComplete = { completed = true },
-    )
-
-    assertTrue(completed)
-  }
-
-  @Test
-  fun `navigateBack from sign up email link does not reroute to pending verification`() {
-    val realBackStack =
-      NavBackStack<NavKey>(
-        AuthDestination.AuthStart,
-        AuthDestination.SignUpEmailLink(emailAddress = "sam@clerk.dev"),
-      )
-    val realAuthState =
-      AuthState(
-        mode = AuthMode.SignInOrUp,
-        backStack = realBackStack,
-        sharedPreferences = preferences(),
-      )
-    val signUp =
-      SignUp(
-        id = "sign_up_123",
-        status = SignUp.Status.MISSING_REQUIREMENTS,
-        requiredFields = listOf("email_address", "password"),
-        optionalFields = emptyList(),
-        missingFields = emptyList(),
-        unverifiedFields = listOf("email_address"),
-        verifications =
-          mapOf(
-            "email_address" to
-              Verification(
-                status = Verification.Status.UNVERIFIED,
-                strategy = Constants.Strategy.EMAIL_LINK,
-              )
-          ),
-        emailAddress = "sam@clerk.dev",
-        passwordEnabled = false,
-      )
-
-    realAuthState.navigateBack()
-    resumeInProgressAuthAttempt(
-      authState = realAuthState,
-      top = realBackStack.lastOrNull(),
-      signIn = null,
-      signUp = signUp,
-      onAuthComplete = {},
-    )
-
-    assertEquals(listOf(AuthDestination.AuthStart), realBackStack.toList())
-  }
-
-  private fun preferences() =
-    context.getSharedPreferences(
-      Constants.Storage.CLERK_PREFERENCES_FILE_NAME,
-      Context.MODE_PRIVATE,
-    )
-}
